refactor(web): replace PostsLoop with PostCards and TechStacksSidebar components (#1513)
- resolved: #1513
