+++ apps/web/src/components/about/my-writings.tsx
@@ -1,49 +1,34 @@
 import React from "react";
-import { Zap } from "lucide-react";
 
-import PostsLoop from "@/components/about/posts-loop";
-import { BlurFadeLi } from "@/components/magicui/blur-fade";
-import { ClickableTechBadges } from "@/lib/tech-badge/utils";
-import config from "@/config";
+import { PostCards } from "@/components/about/post-cards";
+import { TechStacksSidebar } from "@/components/about/tech-stacks-sidebar";
 
 import { BlogPost } from "@/types/post";
-
-import styles from "@/styles/gradient-card.module.css";
-
-import { cn } from "@1chooo/ui/lib/utils";
 
 interface MyWritingsProps {
   count?: number;
   posts?: BlogPost[];
 }
 
-function MyWritings({ count, posts }: MyWritingsProps) {
+function parseDate(dateStr: string): Date {
+  const [month, day, year] = dateStr.split(" ");
+  return new Date(`${month} ${parseInt(day)}, ${year}`);
+}
+
+function MyWritings({ count = 3, posts = [] }: MyWritingsProps) {
+  const sortedPosts = posts
+    .map((post: BlogPost) => ({
+      ...post,
+      date: parseDate(post.publishedAt),
+    }))
+    .sort((a, b) => b.date.getTime() - a.date.getTime())
+    .slice(0, count);
+
   return (
     <div className="w-full max-w-4xl mx-auto my-7 xl:px-0">
       <div className="flex flex-col items-start justify-start md:flex-row md:space-x-7">
-        <PostsLoop count={count} posts={posts} />
-
-        <ul className="w-full mt-10 md:w-1/3 md:mt-0">
-          {Object.entries(config.techStacks).map(([category, badges]) => (
-            <BlurFadeLi inView delay={0.4} direction="up" key={category} className={cn(styles.gradientCard, "mb-4")}>
-              <div className="flex flex-wrap gap-2 shadow-feature-card dark:shadow-feature-card-dark rounded-xl">
-                <div className="relative flex items-center space-x-2">
-                  <Zap className="flex-none text-white-1" size={18} />
-                  <h2 className="flex text-sm font-semibold text-white-1">
-                    {category}
-                  </h2>
-                </div>
-                <div className="flex flex-wrap gap-2 shadow-feature-card dark:shadow-feature-card-dark rounded-xl mt-4">
-                  {badges.map((badgeKey) => (
-                    <span key={badgeKey}>
-                      {ClickableTechBadges[badgeKey]}
-                    </span>
-                  ))}
-                </div>
-              </div>
-            </BlurFadeLi>
-          ))}
-        </ul>
+        <PostCards posts={sortedPosts} />
+        <TechStacksSidebar />
       </div>
     </div>
   );
